# implementation of fbgemm op - regroup_keyed_tensor